🚚 Rename the project
