+++ .pre-commit-config.yaml
@@ -7,7 +7,7 @@ repos:
         language: system
         types: [python]
         entry: poetry run bandit
-        files: ^p1_monitor/
+        files: ^p1monitor/
         require_serial: true
       - id: black
         name: ☕️ Format using black
